+++ src/logger.ts
@@ -1,18 +1,4 @@
-import * as fs from "fs";
-import * as path from "path";
 import * as winston from "winston";
-
-
-/**
- * 获取软件的版本号。
- * 
- * 版本号是形如“major.minor.patch”的字符串。
- */
-export function getAppVersion(): string {
-    const packageJSONPath = path.resolve(`${__dirname}/../package.json`);
-    const packageJSON = JSON.parse(fs.readFileSync(packageJSONPath, "utf8"));
-    return packageJSON.version
-}
 
 
 /**
@@ -40,4 +26,4 @@ logger.add(new winston.transports.Console({
     )
 }));
 
-export { logger };
+export default logger;
